+++ CHANGELOG.md
@@ -9,6 +9,8 @@ We used to have incremental versioning before `0.1.0`.
 
 - Fixes that `_` was marked as invalid by `VagueImportViolation`
 - Fixes that docs for `VagueImportViolation` were misleading
+- Fixes invalid docs for `BracketBlankLineViolation` #1020
+- Add more complex example to `ParametersIndentationViolation` #1021
 
 
 ## 0.13.0 aka The Lintoberfest